refactor: move classes to a new cockpit package
Existing command handlers & producers and CockpitCommandService are moved into new packages
