+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/cockpit/command/CockpitCommandServiceTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service.cockpitcommand;
+package io.gravitee.rest.api.service.cockpit.command;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
