+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/cockpit/command/handler/GoodbyeCommandHandlerTest.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service.impl.commands;
+package io.gravitee.rest.api.service.cockpit.command.handler;
 
-import static io.gravitee.rest.api.service.impl.commands.GoodbyeCommandHandler.DELETED_STATUS;
+import static io.gravitee.rest.api.service.cockpit.command.handler.GoodbyeCommandHandler.DELETED_STATUS;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.*;
