+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/cockpit/command/handler/InstallationCommandHandlerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service.impl.commands;
+package io.gravitee.rest.api.service.cockpit.command.handler;
 
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.Mockito.*;
